+++ tests/test_targets.py
@@ -4,7 +4,20 @@ import pytest
 
 from baybe.exceptions import IncompatibilityError
 from baybe.targets.numerical import NumericalTarget
+from baybe.transformations.core import AffineTransformation
 from baybe.utils.interval import Interval
+
+
+def test_target_addition():
+    """Target addition appends a shifting transformation."""
+    t1 = NumericalTarget("t") + 1
+    assert t1._transformation == AffineTransformation(shift=1)
+
+
+def test_target_multiplication():
+    """Target multiplication appends a scaling transformation."""
+    t1 = NumericalTarget("t") * 2
+    assert t1._transformation == AffineTransformation(scale=2)
 
 
 def test_target_inversion():