+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `total_transformation` and `is_normalized` properties to `NumericalTarget`
 - `invert`, `normalize`, `abs`, `clamp`, `log`, `exp` and `power` methods to
   `NumericalTarget` for easy creation of transformed targets from existing ones
+- Addition and multiplication dunder methods (for scalar values) to `NumericalTarget`
 - `get_image` method to `NumericalTarget` for computing the images of transformed
   target value ranges
 - Support for non-normalized targets in `DesirabilityObjective`